keyset/webview_app/requirements.txt
pywebview==5.1
playwright==1.45.0
requests>=2.31.0
